refactor(page-builder): extract defaults and hooks
