+++ packages/ui/src/components/cms/page-builder/PageBuilder.tsx
@@ -15,13 +15,6 @@ import { Button } from "../../atoms/shadcn";
 import { Toast, Spinner } from "../../atoms";
 import { CheckIcon } from "@radix-ui/react-icons";
 import Palette from "./Palette";
-import {
-  atomRegistry,
-  moleculeRegistry,
-  organismRegistry,
-  containerRegistry,
-  layoutRegistry,
-} from "../blocks";
 import { getShopFromPath } from "@platform-core/utils";
 import useFileUpload from "@ui/hooks/useFileUpload";
 import usePageBuilderState from "./hooks/usePageBuilderState";
@@ -29,54 +22,7 @@ import usePageBuilderDnD from "./hooks/usePageBuilderDnD";
 import PageToolbar from "./PageToolbar";
 import PageCanvas from "./PageCanvas";
 import PageSidebar from "./PageSidebar";
-
-/* ════════════════ component catalogue ════════════════ */
-type ComponentType =
-  | keyof typeof atomRegistry
-  | keyof typeof moleculeRegistry
-  | keyof typeof organismRegistry
-  | keyof typeof containerRegistry
-  | keyof typeof layoutRegistry;
-
-const CONTAINER_TYPES = Object.keys(containerRegistry) as ComponentType[];
-
-const defaults: Partial<Record<ComponentType, Partial<PageComponent>>> = {
-  HeroBanner: { minItems: 1, maxItems: 5 },
-  ValueProps: { minItems: 1, maxItems: 6 },
-  ReviewsCarousel: { minItems: 1, maxItems: 10 },
-  SearchBar: { placeholder: "Search products…", limit: 5 },
-  ProductFilter: { showSize: true, showColor: true, showPrice: true },
-  ProductGrid: {
-    minItems: 1,
-    maxItems: 3,
-    desktopItems: 3,
-    tabletItems: 2,
-    mobileItems: 1,
-    mode: "collection",
-  },
-  ProductCarousel: {
-    minItems: 1,
-    maxItems: 10,
-    desktopItems: 4,
-    tabletItems: 2,
-    mobileItems: 1,
-    mode: "collection",
-  },
-  RecommendationCarousel: { minItems: 1, maxItems: 4 },
-  Testimonials: { minItems: 1, maxItems: 10 },
-  TestimonialSlider: { minItems: 1, maxItems: 10 },
-  ImageSlider: { minItems: 1, maxItems: 10 },
-  AnnouncementBar: {
-    position: "absolute",
-    top: "0",
-    left: "0",
-    width: "100%",
-  },
-  Lookbook: { minItems: 0, maxItems: 10 },
-  MultiColumn: { columns: 2, gap: "1rem" },
-  Divider: { width: "100%", height: "1px" },
-  Spacer: { width: "100%", height: "1rem" },
-};
+import { defaults, CONTAINER_TYPES } from "./defaults";
 
 interface Props {
   page: Page;
